requirements.txt
fastapi
uvicorn
openai
supabase
python-docx
google-api-python-client
google-auth
google-auth-oauthlib
requests
